# openapi/internal.yaml
openapi: 3.0.3
info:
  title: DC Inventory internal API
  version: 0.0.0
...
paths:
...
  /internal/suppliers:
    get:
      operationId: listInternalSuppliers
      summary: List suppliers
      tags:
        - internal-suppliers
      parameters:
        - schema:
            type: string
          in: query
          name: q
          required: false
        - schema:
            default: 1
            type: integer
            minimum: 1
            maximum: 9007199254740991
          in: query
          name: page
          required: false
        - schema:
            default: 25
            type: integer
            minimum: 1
            maximum: 100
          in: query
          name: pageSize
          required: false
        - schema:
            default: vendorNumber
            type: string
            enum:
              - vendorNumber
              - name
          in: query
          name: sortBy
          required: false
        - schema:
            default: asc
            type: string
            enum:
              - asc
              - desc
          in: query
          name: sortOrder
          required: false
      x-table:
        rowId: id
        columns:
          - field: vendorNumber
            label: "Vendor #"
          - field: name
            label: Name
        search:
          param: q
          fields:
            - vendorNumber
            - name
          placeholder: "Search vendor # or name"
        filters: []
        sort:
          defaultBy: vendorNumber
          defaultOrder: asc
          fields:
            - vendorNumber
            - name
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        vendorNumber:
                          type: string
                        name:
                          type: string
                      required:
                        - id
                        - vendorNumber
                        - name
                      additionalProperties: false
                  page:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  pageSize:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                  total:
                    type: integer
                    minimum: -9007199254740991
                    maximum: 9007199254740991
                required:
                  - items
                  - page
                  - pageSize
                  - total
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - invalid_request
                  message:
                    type: string
                    enum:
                      - The request is invalid.
                  requestId:
                    type: string
                required:
                  - error
                  - message
                  - requestId
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - feature_disabled
                required:
                  - error
                additionalProperties: false
    post:
      operationId: createInternalSupplier
      summary: Create supplier
      tags:
        - internal-suppliers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                vendorNumber:
                  type: string
                  minLength: 1
              required:
                - name
                - vendorNumber
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  vendorNumber:
                    type: string
                  name:
                    type: string
                required:
                  - id
                  - vendorNumber
                  - name
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - invalid
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - forbidden
                    required:
                      - error
                    additionalProperties: false
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - feature_disabled
                    required:
                      - error
                    additionalProperties: false
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - duplicate_vendor_number
                required:
                  - error
                additionalProperties: false
  /internal/suppliers/{id}:
    get:
      operationId: getInternalSupplier
      summary: Get supplier
      tags:
        - internal-suppliers
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  vendorNumber:
                    type: string
                  name:
                    type: string
                required:
                  - id
                  - vendorNumber
                  - name
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - invalid
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - feature_disabled
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - not_found
                required:
                  - error
                additionalProperties: false
    patch:
      operationId: updateInternalSupplier
      summary: Update supplier
      tags:
        - internal-suppliers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                vendorNumber:
                  type: string
                  minLength: 1
      parameters:
        - schema:
            type: string
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                    pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                  vendorNumber:
                    type: string
                  name:
                    type: string
                required:
                  - id
                  - vendorNumber
                  - name
                additionalProperties: false
        "400":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - invalid
                required:
                  - error
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - unauthorized
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                anyOf:
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - forbidden
                    required:
                      - error
                    additionalProperties: false
                  - type: object
                    properties:
                      error:
                        type: string
                        enum:
                          - feature_disabled
                    required:
                      - error
                    additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - not_found
                required:
                  - error
                additionalProperties: false
        "409":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    enum:
                      - duplicate_vendor_number
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: http://localhost:3001